pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "deja"
version = "0.2.0"
description = "A personal AI agent that observes your digital life and maintains a living wiki about the people and projects that matter to you"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "google-genai>=1.0.0",
    "google-auth-oauthlib>=1.0.0",
    "Pillow>=10.0",
    "imagehash>=4.3",
    "fastapi>=0.100",
    "uvicorn>=0.20",
    "PyYAML>=6.0",
    "pyobjc-framework-Cocoa>=10.0",
    "mcp>=1.0",
    "httpx>=0.28",
    "mlx-vlm>=0.1",
    "huggingface-hub>=0.24",
    "sqlite-vec>=0.1",
]

[project.scripts]
deja = "deja.__main__:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
deja = [
    "default_assets/CLAUDE.md",
    "default_assets/prompts/*.md",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
markers = [
    "vision: live-network regression tests that call Gemini vision (slow, costs tokens)",
    "live_gws: live gws CLI tests — require Google Workspace auth, no network mocking",
    "real_wiki: opt out of the isolated_home fixture — run against the real wiki",
    "real_awake: opt out of the signal-health autouse is_awake stub",
]
# Exclude vision + live_gws tests from default runs — they hit the network
# and require auth. Run explicitly:
#   pytest -m vision
#   pytest -m live_gws
addopts = "-m 'not vision and not live_gws'"
